+++ packages/atomic/src/components/common/generated-answer/generated-answer-controller.ts
@@ -11,6 +11,7 @@ import {
   StorageItems,
 } from '@/src/utils/local-storage-utils';
 import type {AtomicGeneratedAnswerFeedbackModal} from '../atomic-generated-answer-feedback-modal/atomic-generated-answer-feedback-modal';
+import '@/src/components/common/atomic-generated-answer-feedback-modal/atomic-generated-answer-feedback-modal';
 
 export interface GeneratedAnswerControllerOptions {
   withToggle?: boolean;